refactor(db-client-ts): move promise-queue worker ops into   30_async_operation_worker_v2 (#130)
  Consolidate claimPendingPromiseEventsForWorkers (from
  fsm-instance.ts) and
  ensurePromiseQueueForWorker (from queue.ts), plus the
  PromiseWorkerIdentity /
  EnsurePromiseQueueForWorkerResult types they own, into a new
  src/30_async_operation_worker_v2/asyncOperationWorker.ts —
  mirroring the
  25_async_operation_worker_v1 grouping already applied to the
  v1 async-op
  files.

  Drop now-unused FSM_SCHEMA/FSM_SCHEMA_FN_VERSION imports from
  queue.ts and
  update index.ts's barrel exports and CLAUDE.md accordingly.
  Downstream
  consumers (fsm-core-async-op-worker) import both functions
  through the
  @pgfsm/db package barrel, so no call-site changes are needed
  there.
